http://issues.apache.org/activemq/browse/AMQCPP-30
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk@509365 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-cpp/src/main/activemq/connector/openwire/commands/ActiveMQTextMessage.cpp
@@ -73,7 +73,7 @@ void ActiveMQTextMessage::setText( const std::string& msg ) throw( cms::CMSExcep
         io::ByteArrayOutputStream bos( content );
         io::DataOutputStream dos( &bos );
 
-        dos.writeInt( msg.length() );
+        dos.writeInt( (int)msg.length() );
         dos.write( (const unsigned char*)msg.c_str(), msg.length() );
     }
     AMQ_CATCH_RETHROW( exceptions::ActiveMQException )
